feat: add external Redis configuration and refactor related imports
