+++ apps/dokploy/__test__/queues/redis-connection.test.ts
@@ -1,4 +1,4 @@
-﻿import { afterEach, describe, expect, it, vi } from "vitest";
+import { afterEach, describe, expect, it, vi } from "vitest";
 import fs from "node:fs";
 
 // Mock fs for readSecret tests
