+++ apps/dokploy/server/queues/redis-connection.ts
@@ -1,4 +1,4 @@
-import { redisConfig as sharedRedisConfig } from "@dokploy/server";
+import { redisConfig as sharedRedisConfig } from "@dokploy/server/db/redis-constants";
 import type { ConnectionOptions } from "bullmq";
 
 export const redisConfig = sharedRedisConfig as ConnectionOptions;
