+++ apps/schedules/src/queue.ts
@@ -1,5 +1,5 @@
 import { Queue, type RepeatableJob } from "bullmq";
-import { redisConfig } from "@dokploy/server";
+import { redisConfig } from "@dokploy/server/db/redis-constants";
 import { logger } from "./logger.js";
 import type { QueueJob } from "./schema.js";
