+++ apps/schedules/src/workers.ts
@@ -1,5 +1,5 @@
 import { type Job, Worker } from "bullmq";
-import { redisConfig } from "@dokploy/server";
+import { redisConfig } from "@dokploy/server/db/redis-constants";
 import { logger } from "./logger.js";
 import type { QueueJob } from "./schema.js";
 import { runJobs } from "./utils.js";
