+++ packages/server/src/setup/redis-setup.ts
@@ -1,6 +1,10 @@
 import type { CreateServiceOptions } from "dockerode";
 import { docker } from "../constants";
-import { REDIS_PASSWORD, REDIS_PASSWORD_FILE, readSecret } from "../db/constants";
+import {
+	REDIS_PASSWORD,
+	REDIS_PASSWORD_FILE,
+	readSecret,
+} from "../db/redis-constants";
 import { pullImage } from "../utils/docker/utils";
 
 export const initializeRedis = async () => {
